Create missing-numbers.cpp
this solution is based on CACHING.
Method1: uses maps (can be made more efficient using hash maps). store freq of no. in second list and first list. now match the freq to check which doesn't match
Method2: is faster, this uses array (of size 200). since the range is from 0 to 100.  freq are stored in 100 index value-first_val. when iterating over 1st array increment, when iterating second array decrement. index at which the value is not zero, get corresponding value.